+++ pkg-r/NEWS.md
@@ -3,3 +3,17 @@
 * Initial CRAN submission.
 
 * Added `prompt_template` support for `querychat_system_prompt()`. (Thank you, @oacar! #37, #45)
+
+* `querychat_init()` now accepts a `client`, replacing the previous `create_chat_func` argument. (#60)
+
+  The `client` can be:
+
+  * an `ellmer::Chat` object,
+  * a function that returns an `ellmer::Chat` object,
+  * or a provider-model string, e.g. `"openai/gpt-4.1"`, to be passed to `ellmer::chat()`.
+
+  If `client` is not provided, querychat will use
+
+  * the `querychat.client` R option, which can be any of the above options,
+  * the `QUERYCHAT_CLIENT` environment variable, which should be a provider-model string,
+  * or the default model from `ellmer::chat_openai()`.